server/requirements.txt
Flask==2.1.3
python-dotenv==0.20.0
requests==2.28.1
sib_api_v3_sdk==7.5.0
gunicorn==20.1.0